EmptyResultViewをpublicクラス化
